Fix undefined behavior in char functions
Turns out that passing a signed char to these functions might result in
undefined behavior because the char would be sign extended to integer.

https://stackoverflow.com/a/21754657

The reason for this is kind of historical, but the jist of it is that
the char length was not enough to represent all characters plus the EOF
signal. Therefore they widened the char to int.

https://stackoverflow.com/a/9307359